add a new pseudo project for livecd-rootfs-neon-bionic
since lp doesn't support branches we need a way to address the bionic
version of livecd-rootfs. this is fairly non-trivial since the tooling
Project class doesn't support forcing a packaging scm url (which is a
problem tbh) and instead expects name+component+url_base to construct a
valid packaging scm url. on top of that the launchpad factory attempts to
derive a suitable name from the id in our configs. all in all there's a lot
of assumptions being made that makes it extremely difficult to use this
reasonably.

i.e. ideally we'd want to have the launchpad.net format be an 'id'
and then as a hash to pass for example the *final* url. since that is
a fairly invasive change we'll instead opt to set the packaging scm via an
override for now.

we define a project livecd-rootfs-neon-bionic which will result in a
packaging scm url lp:livecd-rootfs-neon-bionic. that is however not the
valid url and instead we'll override it in the bionic rules to the actually
correct url. additionally we'll restrict the project to bionic only since
it only ever is applicable to bionic due to how bzr has no concept of
branches