move sip4 and pyqt5 to forks
pyqt5 is more like a fork in that we constantly rebuild it because it is
tightly bound to the specific qt version in use. as such it makes no
sense to have it as a backport as backports would get abandoned when
we move series. that's just never the case with pyqt5 (and sip4 by
extension as a tightly related dep)