bring back qtfeedback
was removed in a recent comment as part of the migration to kde qt repos.
qtfeedback is special in that it isn't really being released and hasn't seen
any updates in years. until Bhushan figures out how to proceed with it let's
try to build (not sure that will actually succeed though)